+++ user-api/src/main/kotlin/com/sns/user/component/user/repositories/FriendRequestCrudRepository.kt
@@ -4,6 +4,6 @@ import com.sns.user.component.user.domains.FriendRequest
 import java.util.*
 import org.springframework.data.repository.CrudRepository
 
-interface FriendRequestCrudRepository : CrudRepository<FriendRequest, Int> {
+interface FriendRequestCrudRepository : CrudRepository<FriendRequest, Long> {
     fun findByRequesterIdAndReceiverId(requesterId: String, receiverId: String): Optional<FriendRequest>
 }
